Moved all tests for join functions to a seperate test file.
